i64 is enough 4.838ms for part1 56.59ms for part2

+++ src/int_code.rs
@@ -21,7 +21,7 @@ pub enum Opcode {
     Err,
 }
 
-pub type ValueType = i128;
+pub type ValueType = i64;
 type Memory = Vec<ValueType>;
 
 #[derive(Debug, Clone, Copy)]
